+++ test/fake_cloud_firestore_test.dart
@@ -12,6 +12,7 @@ final from = (DocumentSnapshot<Map<String, dynamic>> snapshot, _) =>
     snapshot.exists ? (Movie()..title = snapshot['title']) : null;
 final to = (Movie? movie, _) =>
     movie == null ? <String, Object?>{} : {'title': movie.title};
+
 void main() {
   group('dump', () {
     const expectedDumpAfterset = '''{
@@ -28,6 +29,20 @@ void main() {
         'name': 'Bob',
       });
       expect(instance.dump(), equals(expectedDumpAfterset));
+    });
+    test('Updates docs', () async {
+      final instance = FakeFirebaseFirestore();
+      final doc = instance.collection('users').doc(uid);
+      await doc.set({'name': 'Bob'});
+      await doc.update({'name': 'Chris'});
+      expect((await doc.get()).get('name'), equals('Chris'));
+    });
+    test('Update fails on non-existent docs', () async {
+      final instance = FakeFirebaseFirestore();
+      expect(
+        instance.collection('messages').doc('newID').update({'name': 'A'}),
+        throwsA(isA<FirebaseException>()),
+      );
     });
     test('Add adds data', () async {
       final instance = FakeFirebaseFirestore();
@@ -600,6 +615,54 @@ void main() {
       expect(snapshot.get('previously absent'), [8, 9]);
     });
 
+    test('FieldValue.arrayUnion() should add unique Maps', () async {
+      final firestore = FakeFirebaseFirestore();
+      final docRef = firestore.collection('test').doc(uid);
+      await docRef.set({
+        'maps': [
+          {'a': 1},
+          {'b': 2}
+        ],
+      });
+
+      await docRef.update({
+        'maps': FieldValue.arrayUnion([
+          {'a': 1},
+          {'c': 3}
+        ]),
+      });
+
+      final snapshot = await docRef.get();
+      final maps = snapshot.get('maps');
+
+      expect(maps, [
+        {'a': 1},
+        {'b': 2},
+        {'c': 3}
+      ]);
+    });
+
+    test('FieldValue.arrayUnion() should add unique dates', () async {
+      final firestore = FakeFirebaseFirestore();
+      final docRef = firestore.collection('test').doc(uid);
+      await docRef.set({
+        'dates': [DateTime(2022, 1, 1)],
+      });
+
+      await docRef.update({
+        'dates':
+            FieldValue.arrayUnion([DateTime(2022, 1, 1), DateTime(2022, 1, 2)]),
+      });
+
+      final snapshot = await docRef.get();
+      final dates = (snapshot.get('dates') as List<dynamic>)
+          .cast<Timestamp>()
+          .map((e) => e.toDate())
+          .toList();
+
+      expect(dates, [DateTime(2022, 1, 1), DateTime(2022, 1, 2)]);
+    });
+
     test('FieldValue.arrayRemove() removes items', () async {
       final firestore = FakeFirebaseFirestore();
       // Empty document before update
@@ -627,6 +690,59 @@ void main() {
       expect(snapshot.get('untouched'), [3]);
       expect(snapshot.get('previously String'), []);
       expect(snapshot.get('previously absent'), []);
+    });
+
+    test('FieldValue.arrayRemove() should remove Maps', () async {
+      final firestore = FakeFirebaseFirestore();
+      final docRef = firestore.collection('test').doc(uid);
+      await docRef.set({
+        'maps': [
+          {'a': 1},
+          {'b': 2},
+          {'c': 3},
+        ],
+      });
+
+      await docRef.update({
+        'maps': FieldValue.arrayRemove([
+          {'b': 2},
+          {'d': 4}
+        ]),
+      });
+
+      final snapshot = await docRef.get();
+      final maps = snapshot.get('maps');
+
+      expect(maps, [
+        {'a': 1},
+        {'c': 3}
+      ]);
+    });
+
+    test('FieldValue.arrayRemove() should remove dates', () async {
+      final firestore = FakeFirebaseFirestore();
+      final docRef = firestore.collection('test').doc(uid);
+      await docRef.set({
+        'dates': [
+          DateTime(2022, 1, 1),
+          DateTime(2022, 1, 2),
+          DateTime(2022, 1, 3),
+        ],
+      });
+
+      await docRef.update({
+        'dates': FieldValue.arrayRemove(
+          [DateTime(2022, 1, 2), DateTime(2022, 1, 4)],
+        ),
+      });
+
+      final snapshot = await docRef.get();
+      final dates = (snapshot.get('dates') as List<dynamic>)
+          .cast<Timestamp>()
+          .map((e) => e.toDate())
+          .toList();
+
+      expect(dates, [DateTime(2022, 1, 1), DateTime(2022, 1, 3)]);
     });
 
     test('FieldValue in nested objects', () async {
@@ -679,6 +795,46 @@ void main() {
       final document = await firestore.doc('root/foo').get();
       // this FieldPath can't be done "a.b" style because the field has dots in it
       expect(document.get(FieldPath(['a', 'I.have.dots'])), 'c');
+    });
+
+    test('Should throw StateError if field does not exist', () async {
+      final firestore = FakeFirebaseFirestore();
+      final collection = firestore.collection('test');
+      final doc = collection.doc('test');
+      await doc.set({
+        'nested': {'field': 3}
+      });
+
+      final snapshot = await doc.get();
+
+      expect(() => snapshot.get('foo'), throwsA(isA<StateError>()));
+      expect(() => snapshot.get('nested.foo'), throwsA(isA<StateError>()));
+    });
+
+    test('Should not throw StateError if value at field is null', () async {
+      final firestore = FakeFirebaseFirestore();
+      final collection = firestore.collection('test');
+      final doc = collection.doc('test');
+      await doc.set({
+        'nested': {'field': null},
+        'field': null
+      });
+
+      final snapshot = await doc.get();
+
+      expect(snapshot.get('field'), isNull);
+      expect(snapshot.get('nested.field'), isNull);
+    });
+
+    test('Should throw StateError for a.b path if a is not Map', () async {
+      final firestore = FakeFirebaseFirestore();
+      final collection = firestore.collection('test');
+      final doc = collection.doc('test');
+      await doc.set({'nested': 3});
+
+      final snapshot = await doc.get();
+
+      expect(() => snapshot.get('nested.field'), throwsA(isA<StateError>()));
     });
   });
 
@@ -862,6 +1018,30 @@ void main() {
     }
   });
 
+  test('Should throw Error if data contains nested list', () async {
+    final firestore = FakeFirebaseFirestore();
+    await expectLater(
+      () => firestore.collection('test').doc().set({
+        'a': [
+          [1, 2],
+          [3, 4]
+        ],
+      }),
+      throwsA(isA<Error>()),
+    );
+    await expectLater(
+      () => firestore.collection('test').doc().set({
+        'a': {
+          'b': [
+            [1, 2],
+            [3, 4]
+          ],
+        },
+      }),
+      throwsA(isA<Error>()),
+    );
+  });
+
   test('auto generate ID', () async {
     final firestore = FakeFirebaseFirestore();
     final reference1 = firestore.collection('users').doc();
@@ -951,7 +1131,7 @@ void main() {
     final foo = firestore.collection('users').doc('foo');
     await foo.set(<String, dynamic>{'name.firstName': 'OldValue Foo'});
     final bar = firestore.collection('users').doc('bar');
-    await foo.set(<String, dynamic>{'name.firstName': 'OldValue Bar'});
+    await bar.set(<String, dynamic>{'name.firstName': 'OldValue Bar'});
 
     final batch = firestore.batch();
     batch.update(foo, <String, dynamic>{'name.firstName': 'Foo'});
@@ -1353,6 +1533,20 @@ void main() {
         expect(snapshot.size, equals(1));
         expect(snapshot.docs.first.data()!.title, equals(MovieTitle));
       }));
+    });
+
+    test('Replace Null with Map', () async {
+      final firestore = FakeFirebaseFirestore();
+
+      var testMap = {'testVal': null};
+      await firestore.collection('testCollection').doc('testDoc').set(testMap);
+
+      var newMap = {
+        'testVal': {'innerKey': 'innerVal'}
+      };
+      expect(
+          firestore.collection('testCollection').doc('testDoc').update(newMap),
+          completes);
     });
   });
 }